requirements.txt
docker==5.0.*
google-cloud-aiplatform==1.1.*